+++ Changelog.md
@@ -4,6 +4,7 @@
 - Save Return to Previous Slot data between reloads
 - Added Return to Previous Slot when depositing items in chests of equipping them
 - Added a way to clear a marked slot when Fake Item is enabled
+- Added Shift-Click to Return to Previous Slot
 
 
 ## v0.9.0.2